requirements.txt
Adyen==3.0.0
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.4.1
authorizenet==1.1.3
Babel==2.8.0
beautifulsoup4==4.7.1
billiard==3.6.4.0
boto3==1.16.9
botocore==1.19.9
braintree==4.4.0
cachetools==4.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==5.1.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.2.8
django-appconf==1.0.4
django-cache-url==3.1.2
django-celery-beat==2.2.1
django-cors-headers==3.9.0
django-countries==7.2.1
django-filter==2.4.0
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-redis==4.12.1
django-render-block==0.7
django-storages==1.10.1
django-templated-email==2.3.0
django-timezone-field==4.2.1
django-versatileimagefield==2.0
djangorestframework==3.12.4
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==4.14.0
flower==1.0.0
freezegun==0.3.15
google-api-core==1.22.2
google-auth==1.21.3
google-cloud-core==1.4.1
google-cloud-pubsub==2.1.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-i18n-address==2.4.0
google-measurement-protocol==1.1.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.13.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
gunicorn==20.0.4
h11==0.9.0
html-to-draftjs==1.0.1
html5lib==1.1
httptools==0.1.1
humanize==3.12.0
idna==2.10
jaeger-client==4.3.0
jdcal==1.4.1
jmespath==0.10.0
jsonfield==3.1.0
kombu==5.1.0
libcst==0.3.10
lxml==4.5.2
Markdown==3.3.3
maxminddb==2.0.3
maxminddb-geolite2==2018.703
measurement==3.2.0
mpmath==1.1.0
mypy-extensions==0.4.3
oauthlib==3.1.0
openpyxl==3.0.5
opentracing==2.3.0
petl==1.6.8
phonenumberslite==8.12.12
Pillow==7.2.0
prices==1.1.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
proto-plus==1.9.1
protobuf==3.13.0
psycopg2-binary==2.9.1
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.9.5
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.14.0
python-json-logger==2.0.1
python-magic==0.4.18
pytimeparse==1.1.8
pytz==2021.3
PyXB==1.2.5
PyYAML==5.3.1
razorpay==1.2.0
redis==3.5.3
requests==2.24.0
rsa==4.6
Rx==1.6.1
s3transfer==0.3.3
sentry-sdk==0.17.3
singledispatch==3.4.0.3
six==1.16.0
soupsieve==2.0.1
sqlparse==0.4.2
stripe==2.55.0
sympy==1.6.2
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.0.2
tornado==6.1
tqdm==4.51.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Unidecode==1.1.1
urllib3==1.25.10
uvicorn==0.12.2
uvloop==0.14.0
vine==5.0.0
watchgod==0.6
wcwidth==0.2.5
WeasyPrint==52
webencodings==0.5.1
websockets==8.1
